Made index integer.
An index for lookup was not explicitly set as integer. This is now
fixed.

+++ pyretis/analysis/flux_analysis.py
@@ -200,7 +200,7 @@ def _calculate_flux(effective_cross, time_in_state, time_window, time_step):
     max_windows = int(1.0 * time_in_state / time_window)
     ncross = np.zeros(max_windows, dtype=np.int)
     for crossing in effective_cross:
-        idx = np.floor((crossing[0] - 0.0) / time_window)
+        idx = int(np.floor((crossing[0] - 0.0) / time_window))
         if idx >= max_windows:
             idx = max_windows - 1
         ncross[idx] += 1
